Extract farmer rpc client into its own module
